+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Support Kotlin 2.4.0-Beta1.
 * Annotations, e.g. `@Ignore`, on functions and classes are correctly propagated.
+* Support configuring the source sets to apply the compiler plugin to.
 
 
 ## [2.13.0-beta1]